Add last.fm API response validation
