+++ lib/utils/lastfm/getTopArtists.ts
@@ -1,0 +1,67 @@
+import { z } from "zod";
+
+import { generateURL } from "./generateURL";
+
+import type {
+  Artist,
+  LastfmParams,
+  Limit,
+  Period,
+  TopArtists,
+  TotalStats,
+} from "@/lib/types/lastfm";
+
+const TopArtistsSchema = z.object({
+  topartists: z.object({
+    artist: z.array(
+      z.object({
+        name: z.string(),
+        playcount: z.string(),
+      })
+    ),
+    "@attr": z.object({ total: z.string() }),
+  }),
+});
+
+/**
+ * @param period - The time period over which to retrieve top artists for.
+ * @param limit - The number of results to fetch. Defaults to 6. Maximum is 10.
+ */
+export async function getTopArtists(
+  period: Period,
+  limit: Limit = 6
+): Promise<TopArtists> {
+  const params: LastfmParams = {
+    method: "user.gettopartists",
+    period,
+    limit,
+  };
+
+  const res = await fetch(generateURL(params), { cache: "no-store" });
+  const { topartists } = TopArtistsSchema.parse(await res.json());
+
+  const artists: TopArtists = topartists.artist.map(
+    (artist): Artist => ({
+      name: artist.name,
+      playcount: artist.playcount,
+    })
+  );
+
+  return artists;
+}
+
+/**
+ * @param period - The time period over which to retrieve top artists for.
+ */
+export async function getTotalArtists(period: Period): Promise<TotalStats> {
+  const params: LastfmParams = {
+    method: "user.gettopartists",
+    period,
+    limit: 1,
+  };
+
+  const res = await fetch(generateURL(params), { cache: "no-store" });
+  const { topartists } = TopArtistsSchema.parse(await res.json());
+
+  return { total: topartists["@attr"].total };
+}
