+++ pages/api/music/recent-track.ts
@@ -1,15 +1,11 @@
-import { getRecentTrack } from "@/utils/lastfm";
+import { getRecentTracks } from "@/utils/lastfm";
 
 export const config = {
   runtime: "edge",
 };
 
 export default async function handler() {
-  const ONE_WEEK_IN_SECONDS = 604_800;
-  const unixTimestamp = Math.floor(Date.now() / 1000); // current Unix timestamp (seconds, 10-digit)
-  const _7DaysAgo = unixTimestamp - ONE_WEEK_IN_SECONDS;
-
-  const recentTrack = await getRecentTrack(_7DaysAgo);
+  const recentTrack = await getRecentTracks();
 
   return new Response(JSON.stringify(recentTrack), {
     status: 200,
